Client ID is now just an int.  If users connect once a millisecond for almost 25 days straight then it will eventually wrap.  I think that's ok.
git-svn-id: https://jmonkeyengine.googlecode.com/svn/trunk@7025 75d07b2b-3a1a-0410-a2c5-0572b91ccdca

+++ engine/src/networking/com/jme3/network/base/DefaultServer.java
@@ -36,7 +36,7 @@ import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.*;
 import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicLong;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -58,7 +58,7 @@ public class DefaultServer implements Server
     static Logger log = Logger.getLogger(DefaultServer.class.getName());
     
     private boolean isRunning = false;
-    private AtomicLong nextId = new AtomicLong(0);
+    private AtomicInteger nextId = new AtomicInteger(0);
     private Kernel reliable;
     private KernelAdapter reliableAdapter;
     private Kernel fast;
@@ -335,7 +335,7 @@ public class DefaultServer implements Server
 
     protected class Connection implements HostedConnection
     {
-        private long id;
+        private int id;
         private Endpoint reliable;
         private Endpoint fast;
